requirements.txt
streamlit
langchain
langchain-community
langchain-openai
chromadb
beautifulsoup4
requests
python-dotenv
tiktoken
